requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
flake8==5.0.4
black==22.6.0
pytest==7.0.1
pytest-cov==3.0.0
gunicorn==20.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
Django==3.2.5
django-crispy-forms==1.9.2
django-widget-tweaks==1.4.8
EasyProcess==0.3
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
pymongo==3.11.0
python-stdnum==1.13
pytz==2020.1
PyVirtualDisplay==1.3.2
redis==3.5.3
djangorestframework==3.11.1
Pillow==9.0.0
drf-yasg==1.17.1
psycopg2